Move Timestamp back over to batch. Fix a type issue not detected in 2.9.3
