+++ summingbird-batch/src/main/scala/com/twitter/summingbird/batch/BatchID.scala
@@ -16,7 +16,6 @@ limitations under the License.
 
 package com.twitter.summingbird.batch
 
-import com.twitter.summingbird.Timestamp
 import com.twitter.algebird.Monoid
 import com.twitter.algebird.{ Universe, Empty, Interval, Intersection,
   InclusiveLower, ExclusiveUpper, InclusiveUpper, ExclusiveLower }
