+++ summingbird-batch/src/main/scala/com/twitter/summingbird/batch/CalendarBatcher.scala
@@ -17,7 +17,6 @@ limitations under the License.
 package com.twitter.summingbird.batch
 
 import java.util.{Calendar, Date, TimeZone}
-import com.twitter.summingbird.Timestamp
 
 object CalendarBatcher {
   /** Type wrapper for the Int's defined in java.util.Calendar */
