+++ summingbird-batch/src/main/scala/com/twitter/summingbird/batch/PrunedSpace.scala
@@ -15,7 +15,6 @@ limitations under the License.
 */
 
 package com.twitter.summingbird.batch
-import com.twitter.summingbird.Timestamp
 
 trait PrunedSpace[-T] extends java.io.Serializable {
   // expire (REMOVE) before writing, T is often (K, V) pair
