+++ summingbird-batch/src/test/scala/com/twitter/summingbird/batch/Generators.scala
@@ -19,7 +19,6 @@ package com.twitter.summingbird.batch
 import org.scalacheck.Arbitrary
 import org.scalacheck.Gen
 
-import com.twitter.summingbird.Timestamp
 import com.twitter.algebird.Interval
 
 /**
