+++ summingbird-chill/src/main/scala/com/twitter/summingbird/chill/SBChillRegistrar.scala
@@ -15,14 +15,13 @@ limitations under the License.
 */
 
 package com.twitter.summingbird.chill
-import com.twitter.summingbird.{MutableStringConfig, SummingbirdConfig, Timestamp}
+import com.twitter.summingbird.{MutableStringConfig, SummingbirdConfig}
 import com.twitter.chill.{ScalaKryoInstantiator, IKryoRegistrar, Kryo, toRich, ReflectingRegistrar, InjectionDefaultRegistrar, InjectionRegistrar}
 import com.twitter.chill.java.IterableRegistrar
 import com.twitter.bijection.Codec
 import com.twitter.chill._
 import com.twitter.chill.config.{ ConfiguredInstantiator => ConfInst }
-import com.twitter.summingbird.batch.BatchID
-
+import com.twitter.summingbird.batch.{BatchID, Timestamp}
 
 /**
  * @author Oscar Boykin
