+++ summingbird-chill/src/main/scala/com/twitter/summingbird/chill/Serializers.scala
@@ -16,9 +16,7 @@ limitations under the License.
 
 package com.twitter.summingbird.chill
 import com.twitter.chill._
-import com.twitter.summingbird.batch.BatchID
-import com.twitter.summingbird.Timestamp
-
+import com.twitter.summingbird.batch.{BatchID, Timestamp}
 
 class BatchIDSerializer extends KSerializer[BatchID] {
   val optimizeForPositive = true
