+++ summingbird-chill/src/test/scala/com/twitter/summingbird/chill/SerializerLaws.scala
@@ -21,7 +21,6 @@ import org.scalacheck.Prop._
 import org.specs2.mutable._
 import com.twitter.chill._
 
-import com.twitter.summingbird.Timestamp
 import com.twitter.summingbird.batch._
 
 object SerializationLaws extends Properties("SerializationLaws") {
