+++ summingbird-scalding/src/main/scala/com/twitter/summingbird/scalding/package.scala
@@ -24,6 +24,7 @@ import cascading.flow.FlowDef
 import com.twitter.algebird.monad.{Reader, StateWithError}
 import com.twitter.algebird.Interval
 import com.twitter.summingbird
+import com.twitter.summingbird.batch.Timestamp
 import org.apache.hadoop.io.Writable
 
 package object scalding {
