+++ summingbird-scalding/src/main/scala/com/twitter/summingbird/scalding/service/BatchedDeltaService.scala
@@ -19,9 +19,8 @@ package com.twitter.summingbird.scalding.service
 import com.twitter.algebird.monad.{StateWithError, Reader}
 import com.twitter.algebird.Semigroup
 import com.twitter.scalding.{Mode, TypedPipe}
-import com.twitter.summingbird.batch.{ BatchID, Batcher }
+import com.twitter.summingbird.batch.{ BatchID, Batcher, Timestamp }
 import cascading.flow.FlowDef
-import com.twitter.summingbird.Timestamp
 import com.twitter.summingbird.scalding._
 
 /**
