+++ summingbird-scalding/src/main/scala/com/twitter/summingbird/scalding/service/SimpleWindowService.scala
@@ -16,8 +16,7 @@
 
 package com.twitter.summingbird.scalding.service
 
-import com.twitter.summingbird.Timestamp
-import com.twitter.summingbird.batch.{ BatchID, Batcher }
+import com.twitter.summingbird.batch.{ BatchID, Batcher, Timestamp }
 import com.twitter.summingbird.scalding._
 import com.twitter.scalding.{Mode, TypedPipe, AbsoluteDuration}
 import com.twitter.algebird.monad.Reader
