+++ summingbird-storm/src/main/scala/com/twitter/summingbird/storm/BaseBolt.scala
@@ -25,7 +25,7 @@ import backtype.storm.tuple.{Tuple, TupleImpl, Fields}
 
 import java.util.{ Map => JMap }
 
-import com.twitter.summingbird.Timestamp
+import com.twitter.summingbird.batch.Timestamp
 import com.twitter.summingbird.storm.option.{AckOnEntry, AnchorTuples, MaxWaitingFutures}
 import com.twitter.summingbird.online.executor.OperationContainer
 import com.twitter.summingbird.online.executor.{InflightTuples, InputState}
